Remove custom mechanicalsoup module. Now project can use original module
