--- mechanicalsoup/__init__.py
@@ -1,8 +1,0 @@
-from .utils import LinkNotFoundError
-from .browser import Browser
-from .form import Form, InvalidFormMethod
-from .stateful_browser import StatefulBrowser
-from .__version__ import __version__
-
-__all__ = ['StatefulBrowser', 'LinkNotFoundError', 'Browser', 'Form',
-           'InvalidFormMethod', '__version__']
